requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
celery==5.6.3
redis==7.4.0
sqlalchemy==2.0.49
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
python-dotenv==1.2.2
